convert the resulting float in set_kmemsize to an int as this needs to be an int when set

+++ nova/virt/openvz_conn.py
@@ -806,10 +806,10 @@ class OpenVzConnection(driver.ComputeDriver):
 
         # Now use the configuration flags to calculate the appropriate
         # values for both barrier and limit.
-        kmem_limit = instance_memory * (
-            float(FLAGS.ovz_kmemsize_percent_of_memory) / 100.0)
-        kmem_barrier = kmem_limit * (
-            float(FLAGS.ovz_kmemsize_barrier_differential) / 100.0)
+        kmem_limit = int(instance_memory * (
+            float(FLAGS.ovz_kmemsize_percent_of_memory) / 100.0))
+        kmem_barrier = int(kmem_limit * (
+            float(FLAGS.ovz_kmemsize_barrier_differential) / 100.0))
         kmemsize = '%d:%d' % (kmem_barrier, kmem_limit)
 
         try:
